# Project/backend/swagger.yaml
openapi: 3.0.0
info:
    title: 'HBL Banking API'
    version: 1.0.0
    description: 'A comprehensive banking API for HBL Clone application'
    contact:
        name: 'Muhammad Junaid Afzal'
        email: me@junaidafzal.dev
    license:
        name: ISC
        url: 'https://opensource.org/licenses/ISC'
servers:
    -
        url: 'http://localhost:8000'
        description: 'Development server'
    -
        url: 'https://your-production-url.com'
        description: 'Production server'
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    schemas:
...
        Notification:
            type: object
            properties: {_id: {type: string}, title: {type: string}, message: {type: string}, type: {type: string}, read: {type: boolean}, createdAt: {type: string, format: date-time}}
...
security:
    -
        bearerAuth: []
paths:
...
    /api/notifications:
        get:
            summary: 'Get user notifications'
            tags: [Notifications]
            security: [{bearerAuth: []}]
            parameters: [{in: query, name: page, schema: {type: integer, default: 1}, description: 'Page number'}, {in: query, name: limit, schema: {type: integer, default: 20}, description: 'Number of notifications per page'}, {in: query, name: type, schema: {type: string}, description: 'Filter by notification type'}]
            responses: {'200': {description: 'List of notifications', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, count: {type: integer}, total: {type: integer}, page: {type: integer}, pages: {type: integer}, data: {type: array, items: {$ref: '#/components/schemas/Notification'}}}}}}}, '401': {description: Unauthorized}}
    '/api/notifications/{id}/read':
        put:
            summary: 'Mark a notification as read'
            tags: [Notifications]
            security: [{bearerAuth: []}]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Notification ID'}]
            responses: {'200': {description: 'Notification marked as read', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, message: {type: string}}}}}}, '401': {description: Unauthorized}, '404': {description: 'Notification not found'}}
    /api/notifications/read-all:
        put:
            summary: 'Mark all notifications as read'
            tags: [Notifications]
            security: [{bearerAuth: []}]
            responses: {'200': {description: 'All notifications marked as read', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, message: {type: string}}}}}}, '401': {description: Unauthorized}}
    '/api/notifications/{id}':
        delete:
            summary: 'Delete a notification'
            tags: [Notifications]
            security: [{bearerAuth: []}]
            parameters: [{in: path, name: id, required: true, schema: {type: string}, description: 'Notification ID'}]
            responses: {'200': {description: 'Notification deleted', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, message: {type: string}}}}}}, '401': {description: Unauthorized}, '404': {description: 'Notification not found'}}
    /api/notifications/unread-count:
        get:
            summary: 'Get unread notification count'
            tags: [Notifications]
            security: [{bearerAuth: []}]
            responses: {'200': {description: 'Unread notification count', content: {application/json: {schema: {type: object, properties: {success: {type: boolean}, data: {type: object, properties: {unreadCount: {type: integer}}}}}}}}, '401': {description: Unauthorized}}
...
    /api/users/notifications:
        put:
            summary: 'Update notification preferences'
            tags: [Users]
            security: [{bearerAuth: []}]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {preferences: {type: object, properties: {notifications: {type: object, properties: {email: {type: boolean, example: true}, sms: {type: boolean, example: false}, push: {type: boolean, example: true}}}}}}}}}}
            responses: {'200': {description: 'Notification preferences updated successfully', content: {application/json: {schema: {type: object, properties: {success: {type: boolean, example: true}, message: {type: string, example: 'Notification preferences updated successfully'}, data: {type: object}}}}}}, '400': {description: 'Validation error'}, '401': {description: Unauthorized}}
...